backend/requirements.txt
# Django Core
Django==5.1
django-filter==23.3
djangorestframework==3.14.0

# Database - PostgreSQL
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.0

# Security & CORS
django-cors-headers==4.3.0

# Environment
python-decouple==3.8

# Static Files
whitenoise==6.6.0

# API Documentation
drf-spectacular==0.27.0

# Production Server
gunicorn==21.2.0

# Image Processing
Pillow==10.1.0

# Essential Dependencies (Eksik olanlar)
setuptools==75.8.0